First 5 problems solutions

+++ python_programs/Practice_Problems/Problem2.py
@@ -1,6 +1,6 @@
 # Write a program that will convert celsius value to fahrenheit
 
-temp_celsius = int(input("Temperature in celsius: "))
+temp_celsius = float(input("Temperature in celsius: "))
 temp_fahrenheit = (temp_celsius * 1.8) + 32
 
 print(f"Temperature in fahrenheit is {temp_fahrenheit}")
